% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Superensemble-matrices.R
\name{array.solve}
\alias{array.solve}
\title{Apply 'solve' over an array}
\usage{
array.solve(arr, solve.over)
}
\arguments{
\item{arr}{Array over which 'solve' is to be applied}

\item{solve.over}{Indices to retain in output}
}
\description{
Applies 'solve' to find a %*% x = b for x, over a multidimensional array, and returns the result as an array of the original dimensions.
}

