% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{box.dot}
\alias{box.dot}
\title{Box density ordinate transform}
\usage{
box.dot(o, mu, s)
}
\arguments{
\item{o}{Vector of observations}

\item{mu}{Vector of fitted means}

\item{s}{Fitted covariance matrix}
}
\value{
Ordinate-transformed value
}
\description{
Run Box density ordinate transform over multivariate normal model to assess calibration.
}
\details{
Returns a centre-outward ordering: outliers tend to have low values, inliers tend to have high values. (Tells us nothing about direction of any bias, though)
}

