% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{det.sharpness}
\alias{det.sharpness}
\title{Determinant sharpness}
\usage{
det.sharpness(sig)
}
\arguments{
\item{sig}{Covariance matrix of fitted model}
}
\value{
Calculated determinant sharpness
}
\description{
Calculate determinant sharpness of a multivariate predictive density.
}

