% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{ensemble.crps}
\alias{ensemble.crps}
\title{Ensemble CRPS}
\usage{
ensemble.crps(model)
}
\arguments{
\item{model}{Array of predictions for which CRPS is to be calculated: variables 路 days 路 years 路 leadtimes 路 ensemble members.}
}
\value{
Array of CRPS scores for each variable and leadtime.
}
\description{
Calculate CRPS for ensemble prediction, for each variable & each leadtime.
}

