% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{ensemble.spread}
\alias{ensemble.spread}
\title{Calculate ensemble spread}
\usage{
ensemble.spread(ens)
}
\arguments{
\item{ens}{Array of ensemble data, variables · days · years · leadtimes · ensemble members}
}
\value{
Array of spreads for each leadtime, variables · leadtimes
}
\description{
For a given forecast ensemble, calculate the spread (standard deviation).
}

