% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{es.crps}
\alias{es.crps}
\title{CRPS for multivariate-normal predictive density}
\usage{
es.crps(o, mu, sig, efc, k = 1000)
}
\arguments{
\item{o}{Vector of observations}

\item{mu}{Vector of means of fitted model}

\item{sig}{Covariance matrix or variance of fitted model - NOT the standard deviation}

\item{efc}{Matrix of multivariate forecasts, each column an ensemble member.}

\item{k}{Number of samples to use in Monte Carlo approximation. Default is 1000.}
}
\value{
Energy score (multivariate CRPS)
}
\description{
CRPS (energy score) for fitted multivariate-normal predictive density.
}
\details{
EITHER parameter 'efc' (for an ensemble forecast) OR parameters 'mu', 'sig' and 'k' (for a fitted univariate or multivariate normal density) must be supplied
}

