% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{fc.mae}
\alias{fc.mae}
\title{Forecast MAE}
\usage{
fc.mae(fc, o)
}
\arguments{
\item{fc}{Vector or array of forecasts to evaluate}

\item{o}{Vector or array of observations. Dimensions must match those of fc}
}
\value{
Calculated MAE
}
\description{
Median Absolute Error calculated over forecasts and observations
}

