% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Superensemble-matrices.R
\name{fit.mimic}
\alias{fit.mimic}
\title{Fit mimic to data}
\usage{
fit.mimic(o, fc, tr, arr.out = T)
}
\arguments{
\item{o}{Observed values of the variables of interest}

\item{fc}{List of ensemble foreasts}

\item{tr}{Matrix of training data}

\item{arr.out}{Logical: return the results in a single array (T - more useful if applying formula over an array) or a list (F - useful if analysing a single forecast). Default is T.}
}
\value{
Either an array or a list containing computed values of Tau and S.
}
\description{
Fit a mimic to an ensemble forecast and training data.
}

