% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{forecast.errors}
\alias{forecast.errors}
\title{Calculate forecast errors}
\usage{
forecast.errors(fc, actual = obs)
}
\arguments{
\item{fc}{Array of forecasts, variables · days · years · leadtimes · ensemble members}

\item{actual}{Array of observations, variables · days · years. Default is object \code{obs}.}
}
\value{
Array of errors, with same dimensions as fc
}
\description{
Get forecast errors against observed values, accounting for forecast leadtime
}

