% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{forecast.rmse}
\alias{forecast.rmse}
\title{Forecast RMSE over full forecast period}
\usage{
forecast.rmse(fc, actual = obs)
}
\arguments{
\item{fc}{Array of forecasts, variables · days · years · leadtimes · ensemble members}

\item{actual}{Array of observations, variables · days · years. Default is object \code{obs}.}
}
\value{
Array of RMSE values, variables · leadtimes · ensemble members
}
\description{
Calculate forecast Root Mean Squared Error against observed values
}

