% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.all}
\alias{load.all}
\title{Import forecasts & obs}
\usage{
load.all()
}
\description{
Function to neatly load required observation & forecast data (keeps code tidy), truncating forecast to cover same date range as observations
}

