% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{load.data}
\alias{load.data}
\title{Load .rda directly into variable}
\usage{
load.data(filenm)
}
\arguments{
\item{filenm}{Name of file (including path) to be loaded}
}
\value{
Loaded object
}
\description{
Wrapper function to load .rda object directly into a named variable, rather than loadng & reassigning
}

