% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{model.crps}
\alias{model.crps}
\title{Calculate fitted CRPS}
\usage{
model.crps(Tau, S)
}
\arguments{
\item{Tau}{Vectors of means for all dimensions}

\item{S}{Variance/covariance matrices for all dimensions}
}
\value{
Array containing CRPS scores for each variable, day, year and leadtime.
}
\description{
Calculate CRPS for fitted model
}

