% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{model.performance}
\alias{model.performance}
\title{Calculate spread, RMSE and CRPS for a fitted model}
\usage{
model.performance(fitted, actual = obs)
}
\arguments{
\item{actual}{Array of observations, against which fitted means are to be assessed. Default is to use built-in \link{obs} data.}

\item{Tau}{Vectors of means for all dimensions}

\item{S}{Variance/covariance matrices for all dimensions}
}
\value{
List containing arrays of calculated spread, RMSE and CRPS
}
\description{
Calculate spread, RMSE and CRPS for a fitted model
}

