% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{model.rmse}
\alias{model.rmse}
\title{Calculate fitted RMSE}
\usage{
model.rmse(Tau, actual = obs)
}
\arguments{
\item{Tau}{Vectors of means for all dimensions}

\item{actual}{Array of observations, against which fitted means are to be assessed. Default is to use built-in \link{obs} data.}
}
\description{
Calculate RMSE for a fitted model
}

