% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{offset.forecast}
\alias{offset.forecast}
\title{Adjust forecast by leadtime}
\usage{
offset.forecast(fc)
}
\arguments{
\item{fc}{Forecast array}
}
\value{
Adjusted forecast array
}
\description{
Takes forecasts at multiple leadtimes and adjusts them so that forecasts for the same day are aligned (eg. look ahead 1 day for t-1, look ahead 2 days for t-2, and so on); rather than aligning records by date on which forecast was produced.
}

