% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{pad.to.5d}
\alias{pad.to.5d}
\title{Pad 4d array to 5d}
\usage{
pad.to.5d(arr)
}
\arguments{
\item{arr}{Array to be padded}
}
\value{
5d array to be used in ensemble processing
}
\description{
Support function - pad an array forecast containing only one ensemble member to behave like an ensemble forecast.
}

