% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dissertation-plots.R
\name{plot.forecast.errors}
\alias{plot.forecast.errors}
\title{Plot forecast errors}
\usage{
plot.forecast.errors(model)
}
\arguments{
\item{model}{Name of dataset to use (use \code{data{package = "SX.weather"}} to list available options)}
}
\description{
Produce a plot of errors for each forecast at each leadtime, for all variables
}

