% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dissertation-plots.R
\name{plot.forecast.rmse}
\alias{plot.forecast.rmse}
\title{Plot RMSE}
\usage{
plot.forecast.rmse(model)
}
\arguments{
\item{model}{Name of dataset to use (use \code{data{package = "SX.weather"}} to list available options)}
}
\description{
Produce a plot of RMSE for each forecast at each leadtime, for all variables
}

