% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dissertation-plots.R
\name{plot.rmse.spread}
\alias{plot.rmse.spread}
\title{Plot RMSE vs spread}
\usage{
plot.rmse.spread(model)
}
\arguments{
\item{model}{Name of dataset to use (use \code{data{package = "SX.weather"}} to list available options)}
}
\description{
Produce a plot of RMSE vs spread for each forecast at each leadtime, for all variables
}

