% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Superensemble-matrices.R
\name{run.model}
\alias{run.model}
\title{Fit model}
\usage{
run.model(varbls = dimnames(obs)[[1]])
}
\arguments{
\item{varbls}{Vector of variables to be included. Default is to include all variables.}
}
\value{
List containing vector of means tau and variance/covariance matrix S
}
\description{
Fit a multivariate normal model to all available data
}

