% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Superensemble-matrices.R
\name{se.covariances}
\alias{se.covariances}
\title{Calculate covariance matrices}
\usage{
se.covariances(fc)
}
\description{
For a superensemble of offset-corrected forecasts, return an array of covariances between perturbed ensemble members for all variables across a single model.
}

