% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Superensemble-matrices.R
\name{se.lambda}
\alias{se.lambda}
\title{Calculate lambda matrix}
\usage{
se.lambda()
}
\description{
For a superensemble of offset-corrected forecasts, return an array of covariances between means of all perturbed ensemble members for all variables across all models.
}

