% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Superensemble-matrices.R
\name{se.precision}
\alias{se.precision}
\title{Calculate precision matrix}
\usage{
se.precision(D, L, vars = 1:dim(D)[[2]])
}
\description{
For a superensemble of offset-corrected forecasts, return array of precision matrices.
}

