% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Superensemble-matrices.R
\name{se.sigma}
\alias{se.sigma}
\title{Calculate sigma matrices}
\usage{
se.sigma(data.list = list(ecmwf, ncep, ukmo))
}
\description{
For a superensemble of offset-corrected forecasts, return an array of covariances between means of all perturbed ensemble members for all variables across all models.
}

