% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{square.mat}
\alias{square.mat}
\title{Square matrix}
\usage{
square.mat(arr)
}
\arguments{
\item{arr}{Array whose first dimension is to be rearranged into a square matrix}
}
\value{
Rearranged array (if possible)
}
\description{
Support function - re-square an array after applying eg. cov
}

