% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{superensemble}
\alias{superensemble}
\title{Create superensemble}
\usage{
superensemble(data.list = list(ecmwf, ncep, ukmo))
}
\arguments{
\item{data.list}{List of model names to be included in superensemble}
}
\value{
Single array containing combined data, with attribute \code{m} a numeric vector labelling source ensemble
}
\description{
Join ensemble forecasts into superensemble, ordered with control forecasts first, followed by perturbations.
}

