% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{u.dev}
\alias{u.dev}
\title{Deviation from uniformity}
\usage{
u.dev(vr, l = max(vr))
}
\arguments{
\item{vr}{Vector of verification ranks, as returned by \code{\link{verif.ranks}}.}

\item{l}{Maximum possible rank. Default is \code{max(vr)}, but if observation is never higher than all ensemble members, this should be manually set to ensemble size + 1.}
}
\value{
Score evaluating degree of deviation from uniformity.
}
\description{
Calculate deviation from uniformity of a vector of verification ranks
}

