% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{verif.ranks}
\alias{verif.ranks}
\title{Verification ranks}
\usage{
verif.ranks(o, ens)
}
\arguments{
\item{o}{Vector or matrix of observations}

\item{ens}{Matrix of forecasts from ensemble. Last dimension gives ensemble size, all preceding dimensions must match observation dimensions.}
}
\value{
Ranks of observation within ensemble spread.
}
\description{
Rank of verifying observation against an ensemble forecast
}
\details{
Treats multivariate ensemble forecast as is univariate, by collapsing all dimensions except ensemble size.
}

