% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance-metrics.R
\name{vr.hist}
\alias{vr.hist}
\title{Verification rank histogram}
\usage{
vr.hist(o, ens, breaks = c(0:10)/10, main = "", col = "skyblue", ...)
}
\arguments{
\item{o}{Vector or matrix of observations}

\item{ens}{Matrix of forecasts from ensemble. Last dimension gives ensemble size, all preceding dimensions must match observation dimensions.}
}
\description{
Plots verification rank histogram by calling \code{\link{verif.ranks}}. Treats multivariates forecast as if univariate.
}

